# src/main/resources/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: spring-harness API (M1)
  version: 1.0.0-m1
  description: >-
    Машиночитаемый контракт M1 — подмножество api-contracts v3 (§0–§3.1, §6) и
    openspec/changes/m1-session-core/specs/session-api. Источники истины — эти документы
    и docs/glossary.md; файл замораживается для генерации серверных интерфейсов и тест-клиента.

    Общие правила (api-contracts §0):
    * **Авторизация**: `Authorization: Bearer <Keycloak JWT>` на всех эндпоинтах; JWT обязан
      содержать группу из `harness.security.allowed-groups` (SSO-гейт, claim `groups`).
      Внутри — одно правило: аутентифицированный видит всё и пишет куда угодно (D-41).
    * **Ошибки**: RFC 9457 Problem Details (`application/problem+json`) с обязательным полем
      `code` из каталога (см. ProblemCode); `code` вне каталога — дефект реализации.
      Для 422 — расширение `errors[]` {pointer, rule, message}.
    * `405 method-not-allowed` возможен на любом пути чужим HTTP-методом — в operations
      не перечисляется (в OpenAPI невыразимо); `406 not-acceptable` — на всех ответах
      с JSON-телом; `415 unsupported-media-type` и `413 payload-too-large` — на всех
      эндпоинтах, принимающих тело (лимит `harness.limits.body`, дефолт 1 МБ).
    * **Пагинация**: конверт `{ items: [...], nextCursor? }` с `?cursor=&limit=`;
      потоковые выборки (messages, events) — `?since=<seq>`, интервал `(since, …]`.
    * **PATCH** — JSON Merge Patch (RFC 7396): отсутствующее поле — не менять,
...

...
servers:
...
security:
  - bearerJWT: []
...
paths:
...
  /sessions/{id}/events:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [SessionEvents]
      operationId: streamSessionEvents
      summary: SSE-поток событий сессии
      description: >-
        Поток `text/event-stream` (api-contracts §3.1). Первым при коннекте/реконнекте
        отправляется снапшот `session.status` {runtimeStatus, lastTurnOutcome?}.
        Далее события: `message.created` — новое видимое событие журнала, данные —
        MessageDto (#/components/schemas/MessageDto; SSE `id:` = seq), `session.status` —
        при смене runtimeStatus/lastTurnOutcome, данные — SessionStatusEvent
        (#/components/schemas/SessionStatusEvent), `ping` — комментарий `: ping` каждые 15 с
        (`harness.sse.ping-interval`); в начале потока — `retry: 5000`.
        Продолжение после разрыва: заголовок `Last-Event-ID` приоритетнее `?since=`;
        доставка за интервал (since, …] — без пропусков и дублей.
        Эндпоинт исключается из генерации интерфейсов (шаг 2): контроллер реализуется
        на SseEmitter вручную; сгенерированный метод этой операции не используется.
      parameters:
        - $ref: '#/components/parameters/Since'
        - name: Last-Event-ID
          in: header
          required: false
          description: seq последнего полученного события (стандартный SSE-реконнект); приоритетнее query-параметра since.
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: >-
            Поток событий; семантика и схемы кадров — в description операции. Тело схемой
            не типизируется (SSE-фреймы), генерация тела не производится — сервер отдаёт
            поток напрямую.
          content:
            text/event-stream:
              examples:
                frame:
                  description: Пример кадра message.created (пустая строка — разделитель кадров SSE).
                  value: |-
                    id: 15
                    event: message.created
                    data: {"id":"01ARZ3NDEKTSV4RRFFQ69G5FAV","seq":15,"kind":"USER","author":"ivanov","payload":{"text":"привет"},"createdAt":"2026-09-18T10:15:00Z"}

        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/SessionNotFound' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
components:
  securitySchemes:
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      description: Идентификатор сессии.
      schema:
        type: string
        format: uuid
    Since:
      name: since
      in: query
      required: false
      description: Курсор seq — события с seq > since (интервал (since, …]); 0 или отсутствие — с начала журнала.
      schema:
        type: integer
        format: int64
        minimum: 0
        default: 0
...
  responses:
    Unauthorized:
      description: '401, code=unauthenticated — нет/невалиден Bearer-токен либо группа JWT вне allowed-groups (SSO-гейт).'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    SessionNotFound:
      description: '404, code=session-not-found.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    NotAcceptable:
      description: '406, code=not-acceptable — Accept не содержит поддерживаемого представления (application/json; для SSE-потока — text/event-stream).'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    ValidationFailed:
      description: '422, code=validation-failed — тело/параметры невалидны; errors[] с pointer/rule/message.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    ProblemCode:
      type: string
      enum:
        - validation-failed
        - unauthenticated
        - session-not-found
        - agent-not-found
        - wrong-session-kind
        - payload-too-large
        - method-not-allowed
        - not-acceptable
        - unsupported-media-type
      description: >-
        Каталог кодов ошибок M1 (api-contracts §6). Точное значение для каждого ответа
        зафиксировано в description соответствующего response-компонента.
...
    ValidationErrorItem:
      type: object
      required: [pointer, rule, message]
      properties:
        pointer:
          type: string
          description: JSON Pointer (RFC 6901) невалидного поля.
          examples:
            - /agentKey
        rule:
          type: string
          description: Машинное имя нарушенного правила (например required, minLength).
        message:
          type: string
          description: Человекочитаемое описание (по-русски).
    ProblemDetail:
      type: object
      description: RFC 9457 Problem Details + расширения code/errors (api-contracts §0.2, §6).
      required: [code]
      properties:
        type:
          type: string
          description: URI-тип проблемы; по умолчанию about:blank.
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        code:
          $ref: '#/components/schemas/ProblemCode'
        errors:
          type: array
          description: Только для 422 validation-failed.
          items:
            $ref: '#/components/schemas/ValidationErrorItem'